+++ src/BMaux.h
@@ -23,7 +23,7 @@ double rand01();
 double randrange(int xmin, int xmax);
 double max(double x, double y);
 double min(double x, double y);
-int quicksort(vector<double> & x, vector<int> & tmp_idx, int first, int last);
+int quicksort(vector<float> & x, vector<int> & tmp_idx, int first, int last);
 void permute(int * vector, int s); 
 int overlap(vector<int> & x1, vector<int> & x2); 
 int convert_char_amino(char a);
@@ -31,6 +31,6 @@ int convert_char_nbase(char a);
 int convert_char_epi(char a);
 int convert_char_ising(char a);
 int print_alphabet(char * ctype); 
-int print_frobenius_norms(vector<double> & h, vector< vector<double> > & J, int L, int q, char *normfile, char *parfile); 
+int print_frobenius_norms(vector<float> & h, vector< vector<float> > & J, int L, int q, char *normfile, char *parfile); 
 
 #endif
